+++ Packages/com.quickeye.importables/Tests/Editor/TestSo.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace QuickEye.ImportableAssets.Editor.Tests
+{
+    internal class TestSo : ScriptableObject
+    {
+
+    }
+}